requirements.txt
langchain>=0.1.0
openai>=1.0.0
networkx>=3.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
pydantic>=2.0.0
tqdm>=4.65.0
neo4j>=5.28.0